fix xoshiro allways 0

+++ src/xoshiro256plus.hpp
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-float next(void);
+uint64_t next(void);
 
 #endif
